+++ src/types.ts
@@ -1,11 +1,11 @@
 /**
  * Form data entry type.
  */
-export type FormDataEntry = {
+export interface FormDataEntry {
   path: string;
   input: File | string;
   output: boolean | Date | File | null | number | string | undefined;
-};
+}
 
 /**
  * Form data transform type.
@@ -15,10 +15,10 @@ export type FormDataTransform = (entry: FormDataEntry) => any;
 /**
  * Form data info type.
  */
-export type FormDataInfo = Partial<{
-  arrays: string[];
-  booleans: string[];
-  dates: string[];
-  files: string[];
-  numbers: string[];
-}>;
+export interface FormDataInfo {
+  arrays?: string[];
+  booleans?: string[];
+  dates?: string[];
+  files?: string[];
+  numbers?: string[];
+}